backend_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.0
python-multipart==0.0.9
httpx==0.27.0
geoalchemy2==0.14.5
shapely==2.0.4
aiosmtplib==3.0.1
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.2
